changed: use signed loop variable in OpenMP loop
prior to OpenMP 3.0, unsigned loop variables is unsupported.

+++ opm/input/eclipse/EclipseState/Grid/EclipseGrid.cpp
@@ -66,13 +66,17 @@
 
 #include <algorithm>
 #include <cmath>
+#include <cstddef>
 #include <cstring>
+#include <cstdint>
 #include <functional>
 #include <initializer_list>
 #include <numeric>
 #include <tuple>
 #include <functional>
 #include <unordered_map>
+#include <stdexcept>
+#include <vector>
 
 #include <fmt/format.h>
 
@@ -1555,7 +1559,7 @@ std::vector<double> EclipseGrid::createDVector(const std::array<int,3>& dims, st
             std::vector<double> volume(this->m_nactive);
 
             #pragma omp parallel for schedule(static)
-            for (std::size_t active_index = 0; active_index < this->m_active_to_global.size(); active_index++) {
+            for (std::int64_t active_index = 0; active_index < static_cast<std::int64_t>(this->m_active_to_global.size()); active_index++) {
                 std::array<double,8> X;
                 std::array<double,8> Y;
                 std::array<double,8> Z;
